SalesMCP/requirements.txt
fastapi
uvicorn
psycopg2-binary
pyyaml
pydantic
python-dotenv
langchain
langchain-google-genai
sqlalchemy
pydantic-settings
jinja2
